refactor live user component
